fixed progress for stopped and filtered torrents

+++ lib/Torrent.cc
@@ -80,7 +80,7 @@ Torrent::Torrent(const Torrent::ResumeInfo& ri, bool queued) : m_prop_handle(*th
 	else
 		m_trackers = m_info.trackers();
 
-	for (size_t i = 0; i < m_trackers.size(); i++)
+	for (unsigned int i = 0; i < m_trackers.size(); i++)
 		m_replies[m_trackers[i].url] = "";
 }
 
@@ -350,7 +350,7 @@ void Torrent::set_tracker_reply(const Glib::ustring& reply, const Glib::ustring&
 
 	if (tracker.empty())
 	{
-		for (size_t j = 0; j < trackers.size(); j++)
+		for (unsigned int j = 0; j < trackers.size(); j++)
 		{
 			if (trackers[j].tier == m_cur_tier)
 			{
@@ -452,9 +452,9 @@ void Torrent::add_tracker(const Glib::ustring& url)
 			diff = true;
 		else if (m_trackers.size() == trackers.size())
 		{
-			for (size_t i = 0; i < m_trackers.size() && !diff; i++)
+			for (unsigned int i = 0; i < m_trackers.size() && !diff; i++)
 			{
-				for (size_t j = 0; j < trackers.size(); j++)
+				for (unsigned int j = 0; j < trackers.size(); j++)
 				{
 					if (m_trackers[i].url == trackers[j].url)
 					{
@@ -542,7 +542,7 @@ void Torrent::reannounce(const Glib::ustring& tracker)
 				break;
 			}
 		}
-		for (size_t i = 1; i < trackers.size(); i++)
+		for (unsigned int i = 1; i < trackers.size(); i++)
 		{
 			if (trackers[i].tier < m_cur_tier)
 				trackers[i].tier++;
@@ -612,7 +612,7 @@ const libtorrent::entry Torrent::get_resume_entry(bool stopping, bool quitting)
 	resume_entry["completed"] = m_completed;
 
 	libtorrent::entry::list_type e_filter;
-	for (size_t i = 0; i < m_filter.size(); i++)
+	for (unsigned int i = 0; i < m_filter.size(); i++)
 		if (m_filter[i])
 			e_filter.push_back(i);
 	resume_entry["filter"] = e_filter;
@@ -621,7 +621,7 @@ const libtorrent::entry Torrent::get_resume_entry(bool stopping, bool quitting)
 		resume_entry["group"] = m_group;
 
 	libtorrent::entry::list_type e_trackers;
-	for (size_t i = 0; i < m_trackers.size(); i++)
+	for (unsigned int i = 0; i < m_trackers.size(); i++)
 		e_trackers.push_back(m_trackers[i].url);
 	resume_entry["trackers"] = e_trackers;
